+++ indigo/TestUtils.cpp
@@ -16,7 +16,7 @@ void doTestAssert(bool expr, const char* test, long line, const char* file)
 {
 	if(!expr)
 	{
-		conPrint("Test Assertion Failed: " + std::string(file) + ", line " + toString(line) + ":\n" + std::string(test));
+		conPrint("Test Assertion Failed: " + std::string(file) + ", line " + toString((int)line) + ":\n" + std::string(test));
 		exit(0);
 	}
 }
